extract Google Books results from Suggest menu component
