+++ packages/databyss-editor/components/Suggest/index.js
@@ -1,0 +1,2 @@
+export { default as GoogleBooks } from './GoogleBooks'
+export { default as SuggestMenu } from './SuggestMenu'
